refactor(scripts): move legacy scan orchestrator
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
